Moved pty handling to module
